Move active item related value to a separate provider
